+++ packages/nornj-react/src/bindTemplate.ts
@@ -1,16 +1,16 @@
-import nj from 'nornj';
+import nj, { ElementType } from 'nornj';
 
-function _registerComponent(name: string, component: React.ElementType) {
+function _registerComponent(name: string, component: ElementType) {
   if (name != null) {
     nj.registerComponent(name, component);
   }
 }
 
-export function bindTemplate<T extends React.ElementType>(target: T): T;
-export function bindTemplate(name: string | React.ElementType): <T extends React.ElementType>(target: T) => T;
+export function bindTemplate<T extends ElementType>(target: T): T;
+export function bindTemplate(name: string | ElementType): <T extends ElementType>(target: T) => T;
 export function bindTemplate(name) {
   if (nj.isString(name)) {
-    return function(component: React.ElementType) {
+    return function(component: ElementType) {
       _registerComponent(name, component);
 
       return component;