+++ packages/nornj-react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [v5.0.1] 2019.12.04
+
+* 🏷️ Change the `React.ElementType` to the `NornJ.ElementType` in type definitions.
+
 ## [v5.0.0] 2019.12.03
 
 * 🚩 Support complete typescript type definition.